Decorator-based Module definition. Dep-Order Metaclass remains, on the side
